# legion/operator/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a EDI server.",
        "title": "EDI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/api/v1/model/deployment": {
            "get": {
                "description": "Get list of Model deployments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelDeployment"
                ],
                "summary": "Get list of Model deployments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model id label",
                        "name": "com.epam.legion.model.id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "model version label",
                        "name": "com.epam.legion.model.version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1.MDResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a Model deployment. Result is updated Model deployment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelDeployment"
                ],
                "summary": "Update a Model deployment",
                "parameters": [
                    {
                        "description": "Update a Model deployment",
                        "name": "md",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.MDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a Model deployment. Result is created Model deployment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelDeployment"
                ],
                "summary": "Create a Model deployment",
                "parameters": [
                    {
                        "description": "Create a Model deployment",
                        "name": "md",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.MDRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete list of Model deployments by labels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelDeployment"
                ],
                "summary": "Delete list of Model deployments by labels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model id label",
                        "name": "com.epam.legion.model.id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "model version label",
                        "name": "com.epam.legion.model.version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            }
        },
        "/api/v1/model/deployment/{name}": {
            "get": {
                "description": "Get a Model deployment by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelDeployment"
                ],
                "summary": "Get a Model deployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model deployment name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.MDResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Model deployment by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelDeployment"
                ],
                "summary": "Delete a Model deployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model deployment name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            }
        },
        "/api/v1/model/deployment/{name}/scale": {
            "put": {
                "description": "Scale a Model deployment. Result is updated Model deployment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelDeployment"
                ],
                "summary": "Scale a Model deployment",
                "parameters": [
                    {
                        "description": "Scale a Model deployment",
                        "name": "md",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.MDReplicas"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Name of Model Deployment",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "routes.HTTPResult": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Success of error message",
                    "type": "string"
                }
            }
        },
        "v1.MDReplicas": {
            "type": "object",
            "properties": {
                "replicas": {
                    "description": "New number of replicas",
                    "type": "integer"
                }
            }
        },
        "v1.MDRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "Model deployment name",
                    "type": "string"
                },
                "spec": {
                    "description": "Model Deployment specification. It is the same as ModelDeployment CRD specification",
                    "type": "object",
                    "$ref": "#/definitions/v1alpha1.ModelDeploymentSpec"
                }
            }
        },
        "v1.MDResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "Model deployment name",
                    "type": "string"
                },
                "spec": {
                    "description": "Model Deployment specification. It is the same as ModelDeployment CRD specification",
                    "type": "object",
                    "$ref": "#/definitions/v1alpha1.ModelDeploymentSpec"
                },
                "status": {
                    "description": "Model Deployment specification. It is the same as ModelDeployment CRD specification",
                    "type": "object",
                    "$ref": "#/definitions/v1alpha1.ModelDeploymentStatus"
                }
            }
        },
...
        "v1alpha1.ModelDeploymentSpec": {
            "type": "object",
            "properties": {
                "annotations": {
                    "description": "Annotations for model pods.",
                    "type": "object"
                },
                "image": {
                    "description": "Model Docker image",
                    "type": "string"
                },
                "livenessProbeInitialDelay": {
                    "description": "Initial delay for liveness probe of model pod",
                    "type": "integer"
                },
                "readinessProbeInitialDelay": {
                    "description": "Initial delay for readiness probe of model pod",
                    "type": "integer"
                },
                "replicas": {
                    "description": "Number of pods for model. By default the replicas parameter equals 1.",
                    "type": "integer"
                },
                "resources": {
                    "description": "Resources for model deployment\nThe same format like k8s uses for pod resources.",
                    "type": "string"
                }
            }
        },
        "v1alpha1.ModelDeploymentStatus": {
            "type": "object",
            "properties": {
                "deployment": {
                    "description": "The model k8s deployment name",
                    "type": "string"
                },
                "message": {
                    "description": "The message describes the state in more details.",
                    "type": "string"
                },
                "service": {
                    "description": "The model k8s service name",
                    "type": "string"
                },
                "serviceURL": {
                    "description": "The model k8s service name",
                    "type": "string"
                },
                "state": {
                    "description": "The state of a model deployment.\n  \"DeploymentFailed\" - A model was not deployed. Because some parameters of the\n                       custom resource are wrong. For example, there is not a model\n                       image in a Docker registry.\n  \"DeploymentCreated\" - A model was deployed successfully.",
                    "type": "string"
                }
            }
        },
...
    }
}